+++ dropbox_backup/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Improved log messages.
+
+### Added
+
+- Rename backup files to the backup name (e.g. 'my_backup.tar' instead of '36tdff.tar').
+
 <!-- markdown-link-check-disable -->
 
 ## [0.5.3](https://github.com/mikevansighem/dropbox-backup/releases/tag/0.5.3) - 2021-12-22